requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
aiofiles==24.1.0
    # via gradio
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   gradio
    #   httpx
    #   starlette
brotli==1.1.0
    # via gradio
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   typer
    #   uvicorn
faiss-cpu==1.11.0.post1
    # via fot-intervention-recommender (pyproject.toml)
fastapi==0.116.1
    # via gradio
ffmpy==0.6.1
    # via gradio
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.7.0
    # via
    #   gradio-client
    #   huggingface-hub
    #   torch
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.25.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.178.0
    # via google-generativeai
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.5
    # via fot-intervention-recommender (pyproject.toml)
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
gradio==5.41.0
    # via fot-intervention-recommender (pyproject.toml)
gradio-client==1.11.0
    # via
    #   fot-intervention-recommender (pyproject.toml)
    #   gradio
greenlet==3.2.3
    # via sqlalchemy
groovy==0.1.2
    # via gradio
grpcio==1.74.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.1.5
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   gradio
    #   gradio-client
    #   langsmith
    #   safehttpx
huggingface-hub==0.34.3
    # via
    #   gradio
    #   gradio-client
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   gradio
    #   torch
joblib==1.5.1
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.27
    # via fot-intervention-recommender (pyproject.toml)
langchain-core==0.3.72
    # via
    #   langchain
    #   langchain-text-splitters
langchain-text-splitters==0.3.9
    # via langchain
langsmith==0.4.10
    # via
    #   langchain
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   gradio
    #   jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==1.26.4
    # via
    #   fot-intervention-recommender (pyproject.toml)
    #   faiss-cpu
    #   gradio
    #   pandas
    #   scikit-learn
    #   scipy
    #   transformers
orjson==3.11.1
    # via
    #   gradio
    #   langsmith
packaging==25.0
    # via
    #   faiss-cpu
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   transformers
pandas==2.3.1
    # via gradio
pillow==11.3.0
    # via
    #   gradio
    #   sentence-transformers
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   fastapi
    #   google-generativeai
    #   gradio
    #   langchain
    #   langchain-core
    #   langsmith
pydantic-core==2.33.2
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.19.2
    # via rich
pyparsing==3.2.3
    # via httplib2
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via fot-intervention-recommender (pyproject.toml)
python-multipart==0.0.20
    # via gradio
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   gradio
    #   huggingface-hub
    #   langchain
    #   langchain-core
    #   transformers
regex==2025.7.34
    # via transformers
requests==2.32.4
    # via
    #   google-api-core
    #   huggingface-hub
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   transformers
requests-toolbelt==1.0.0
    # via langsmith
rich==14.1.0
    # via typer
rsa==4.9.1
    # via google-auth
ruff==0.12.7
    # via gradio
safehttpx==0.1.6
    # via gradio
safetensors==0.5.3
    # via transformers
scikit-learn==1.7.1
    # via sentence-transformers
scipy==1.16.1
    # via
    #   scikit-learn
    #   sentence-transformers
semantic-version==2.10.0
    # via gradio
sentence-transformers==5.0.0
    # via fot-intervention-recommender (pyproject.toml)
setuptools==80.9.0
    # via fot-intervention-recommender (pyproject.toml)
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.42
    # via langchain
starlette==0.47.2
    # via
    #   fastapi
    #   gradio
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via langchain-core
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.4
    # via transformers
tomlkit==0.13.3
    # via gradio
torch==2.2.2
    # via
    #   fot-intervention-recommender (pyproject.toml)
    #   sentence-transformers
tqdm==4.67.1
    # via
    #   google-generativeai
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.54.1
    # via
    #   fot-intervention-recommender (pyproject.toml)
    #   sentence-transformers
typer==0.16.0
    # via gradio
typing-extensions==4.14.1
    # via
    #   anyio
    #   fastapi
    #   google-generativeai
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sentence-transformers
    #   sqlalchemy
    #   starlette
    #   torch
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via gradio
websockets==15.0.1
    # via gradio-client
zstandard==0.23.0
    # via langsmith
